[fix] 템플릿 전부 수정 완료, 캐시 에러 수정

+++ main/src/main/java/codeview/main/businessservice/board/presentation/dao/BoardForm.java
@@ -29,6 +29,6 @@ public class BoardForm {
 
     private List<MultipartFile> boardMultipartFileList;
 
-    private Integer problemId;
+    private Long problemId;
 
 }
